requirements.txt
python-telegram-bot==22.0.0
httpx~=0.28
httpcore==1.*
anyio>=4.0
certifi
idna
sniffio
h11>=0.16
Quart==0.18.3
Werkzeug==2.3.7
aiofiles
blinker
click>=8.0.0
hypercorn>=0.17
itsdangerous
jinja2
markupsafe
nest_asyncio==1.6.0
